src/discord_kg/ingestion/discord_client.py
"""Discord API client for fetching messages"""

import asyncio
from typing import List, Dict, Any
import discord
from discord.ext import commands

class DiscordIngestor:
    """Handles Discord API interactions for message fetching"""
    
    def __init__(self, token: str, server_id: int):
        self.token = token
        self.server_id = server_id
        self.client = None
    
    async def fetch_all_messages(self) -> List[Dict[str, Any]]:
        """Fetch all historical messages from all channels"""
        # TODO: Implement Discord API message fetching
        # - Connect to Discord API
        # - Iterate through all channels
        # - Fetch message history with pagination
        # - Handle rate limits
        # - Return structured message data
        pass
    
    async def fetch_channel_messages(self, channel_id: int) -> List[Dict[str, Any]]:
        """Fetch messages from a specific channel"""
        # TODO: Implement channel-specific message fetching
        pass