src/discord_kg/storage/postgres_client.py
"""PostgreSQL client for metadata and state tracking"""

import psycopg2
from typing import Dict, Any, List

class PostgresClient:
    """PostgreSQL client for processing state and metadata"""
    
    def __init__(self, connection_string: str):
        self.connection_string = connection_string
        self.connection = None
    
    def connect(self):
        """Establish database connection"""
        # TODO: Implement database connection
        pass
    
    def store_processing_state(self, batch_id: str, state: Dict[str, Any]) -> None:
        """Store batch processing state"""
        # TODO: Track which batches have been processed
        pass
    
    def get_processing_state(self, batch_id: str) -> Dict[str, Any]:
        """Get processing state for a batch"""
        # TODO: Retrieve processing state
        pass
    
    def close(self):
        """Close database connection"""
        if self.connection:
            self.connection.close()