pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-kg-pipeline"
version = "0.1.0"
description = "Discord Knowledge Graph Pipeline"
authors = [{name = "Arthur", email = "arthur@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.3.2",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.1",
    "psycopg2-binary>=2.9.9",
    "neo4j>=5.14.1",
    "boto3>=1.34.0",
    "b2sdk>=1.24.1",
    "openai>=1.3.5",
    "langchain>=0.0.350",
    "langchain-community>=0.0.1",
    "langchain-openai>=0.0.2",
    "pandas>=2.1.4",
    "numpy>=1.25.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "click>=8.1.7",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"